pyproject.toml
[project]
name = "ezra-talks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bertopic>=0.17.4",
    "numpy>=2.4.4",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "umap-learn>=0.5.12",
]
